Main <!-- Thank you for submitting a Pull Request. Please: * Read our Pull Request guidelines:   https://github.com/microsoft/vscode/wiki/How-to-Contribute#pull-requests * Associate an issue with the Pull Request. * Ensure that the code is up-to-date with the `main` branch. * Include a description of the proposed changes and how to test them. -->
