+++ src/vs/workbench/contrib/extensions/browser/Extension.ts
@@ -1,0 +1,3 @@
+
+
+
