+++ src/vs/workbench/contrib/extensions/browser/extensionsWorkbenchService.ts
@@ -6,7 +6,8 @@
 import * as nls from 'vs/nls';
 import * as semver from 'vs/base/common/semver/semver';
 import { Event, Emitter } from 'vs/base/common/event';
-import { firstOrDefault, index } from 'vs/base/common/arrays';
+import { firstOrDefault } from 'vs/base/common/
+';
 import { CancelablePromise, Promises, ThrottledDelayer, createCancelablePromise } from 'vs/base/common/async';
 import { CancellationError, isCancellationError } from 'vs/base/common/errors';
 import { Disposable, MutableDisposable, toDisposable } from 'vs/base/common/lifecycle';
@@ -14,12 +15,12 @@ import { IPager, singlePagePager } from 'vs/base/common/paging';
 import { ITelemetryService } from 'vs/platform/telemetry/common/telemetry';
 import {
 	IExtensionGalleryService, ILocalExtension, IGalleryExtension, IQueryOptions,
-	InstallExtensionEvent, DidUninstallExtensionEvent, InstallOperation, WEB_EXTENSION_TAG, InstallExtensionResult,
-	IExtensionsControlManifest, IExtensionInfo, IExtensionQueryOptions, IDeprecationInfo, isTargetPlatformCompatible, InstallExtensionInfo, EXTENSION_IDENTIFIER_REGEX,
+	InstallOperation, WEB_EXTENSION_TAG, InstallExtensionResult,
+	IExtensionsControlManifest, IExtensionInfo, IExtensionQueryOptions, isTargetPlatformCompatible, InstallExtensionInfo, EXTENSION_IDENTIFIER_REGEX,
 	InstallOptions, IProductVersion
 } from 'vs/platform/extensionManagement/common/extensionManagement';
-import { IWorkbenchExtensionEnablementService, EnablementState, IExtensionManagementServerService, IExtensionManagementServer, IWorkbenchExtensionManagementService, DefaultIconPath, IResourceExtension, extensionsConfigurationNodeBase } from 'vs/workbench/services/extensionManagement/common/extensionManagement';
-import { getGalleryExtensionTelemetryData, getLocalExtensionTelemetryData, areSameExtensions, groupByExtension, getGalleryExtensionId } from 'vs/platform/extensionManagement/common/extensionManagementUtil';
+import { IWorkbenchExtensionEnablementService, EnablementState, IExtensionManagementServerService, IExtensionManagementServer, IWorkbenchExtensionManagementService, IResourceExtension, extensionsConfigurationNodeBase } from 'vs/workbench/services/extensionManagement/common/extensionManagement';
+import { areSameExtensions, groupByExtension, getGalleryExtensionId } from 'vs/platform/extensionManagement/common/extensionManagementUtil';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
 import { IHostService } from 'vs/workbench/services/host/browser/host';
@@ -31,14 +32,12 @@ import { ExtensionsInput, IExtensionEditorOptions } from 'vs/workbench/contrib/e
 import { ILogService } from 'vs/platform/log/common/log';
 import { IProgressOptions, IProgressService, ProgressLocation } from 'vs/platform/progress/common/progress';
 import { INotificationService, Severity } from 'vs/platform/notification/common/notification';
-import * as resources from 'vs/base/common/resources';
 import { CancellationToken } from 'vs/base/common/cancellation';
 import { IStorageService, StorageScope, StorageTarget } from 'vs/platform/storage/common/storage';
 import { IFileService } from 'vs/platform/files/common/files';
-import { IExtensionManifest, ExtensionType, IExtension as IPlatformExtension, TargetPlatform, ExtensionIdentifier, IExtensionIdentifier, IExtensionDescription, isApplicationScopedExtension } from 'vs/platform/extensions/common/extensions';
+import { ExtensionType, ExtensionIdentifier, IExtensionIdentifier, IExtensionDescription, isApplicationScopedExtension } from 'vs/platform/extensions/common/extensions';
 import { ILanguageService } from 'vs/editor/common/languages/language';
 import { IProductService } from 'vs/platform/product/common/productService';
-import { FileAccess } from 'vs/base/common/network';
 import { IIgnoredExtensionsManagementService } from 'vs/platform/userDataSync/common/ignoredExtensions';
 import { IUserDataAutoSyncService, IUserDataSyncEnablementService, SyncResource } from 'vs/platform/userDataSync/common/userDataSync';
 import { IContextKey, IContextKeyService } from 'vs/platform/contextkey/common/contextkey';
@@ -57,11 +56,12 @@ import { IUpdateService, StateType } from 'vs/platform/update/common/update';
 import { isEngineValid } from 'vs/platform/extensions/common/extensionValidator';
 import { IUriIdentityService } from 'vs/platform/uriIdentity/common/uriIdentity';
 import { IWorkspaceContextService } from 'vs/platform/workspace/common/workspace';
-import { ShowCurrentReleaseNotesActionId } from 'vs/workbench/contrib/update/common/update';
 import { Registry } from 'vs/platform/registry/common/platform';
 import { IConfigurationRegistry, Extensions as ConfigurationExtensions } from 'vs/platform/configuration/common/configurationRegistry';
+import { Extension } from './abstractRuntimeExtensionsEditor';
+import { Extensions } from './abstractRuntimeExtensionsEditor';
 
-interface IExtensionStateProvider<T> {
+export interface IExtensionStateProvider<T> {
 	(extension: Extension): T;
 }
 
@@ -76,772 +76,7 @@ type ExtensionsLoadClassification = {
 	readonly count: { classification: 'PublicNonPersonalData'; purpose: 'FeatureInsight'; comment: 'The number of extensions that are installed.' };
 };
 
-export class Extension implements IExtension {
-
-	public enablementState: EnablementState = EnablementState.EnabledGlobally;
-	public readonly resourceExtension: IResourceExtension | undefined;
-
-	constructor(
-		private stateProvider: IExtensionStateProvider<ExtensionState>,
-		private runtimeStateProvider: IExtensionStateProvider<ExtensionRuntimeState | undefined>,
-		public readonly server: IExtensionManagementServer | undefined,
-		public local: ILocalExtension | undefined,
-		public gallery: IGalleryExtension | undefined,
-		private readonly resourceExtensionInfo: { resourceExtension: IResourceExtension; isWorkspaceScoped: boolean } | undefined,
-		@IExtensionGalleryService private readonly galleryService: IExtensionGalleryService,
-		@ITelemetryService private readonly telemetryService: ITelemetryService,
-		@ILogService private readonly logService: ILogService,
-		@IFileService private readonly fileService: IFileService,
-		@IProductService private readonly productService: IProductService
-	) {
-		this.resourceExtension = resourceExtensionInfo?.resourceExtension;
-	}
-
-	get type(): ExtensionType {
-		return this.local ? this.local.type : ExtensionType.User;
-	}
-
-	get isBuiltin(): boolean {
-		return this.local ? this.local.isBuiltin : false;
-	}
-
-	get isWorkspaceScoped(): boolean {
-		if (this.local) {
-			return this.local.isWorkspaceScoped;
-		}
-		if (this.resourceExtensionInfo) {
-			return this.resourceExtensionInfo.isWorkspaceScoped;
-		}
-		return false;
-	}
-
-	get name(): string {
-		if (this.gallery) {
-			return this.gallery.name;
-		}
-		return this.getManifestFromLocalOrResource()?.name ?? '';
-	}
-
-	get displayName(): string {
-		if (this.gallery) {
-			return this.gallery.displayName || this.gallery.name;
-		}
-
-		return this.getManifestFromLocalOrResource()?.displayName ?? this.name;
-	}
-
-	get identifier(): IExtensionIdentifier {
-		if (this.gallery) {
-			return this.gallery.identifier;
-		}
-		if (this.resourceExtension) {
-			return this.resourceExtension.identifier;
-		}
-		return this.local!.identifier;
-	}
-
-	get uuid(): string | undefined {
-		return this.gallery ? this.gallery.identifier.uuid : this.local?.identifier.uuid;
-	}
-
-	get publisher(): string {
-		if (this.gallery) {
-			return this.gallery.publisher;
-		}
-		return this.getManifestFromLocalOrResource()?.publisher ?? '';
-	}
-
-	get publisherDisplayName(): string {
-		if (this.gallery) {
-			return this.gallery.publisherDisplayName || this.gallery.publisher;
-		}
-
-		if (this.local?.publisherDisplayName) {
-			return this.local.publisherDisplayName;
-		}
-
-		return this.publisher;
-	}
-
-	get publisherUrl(): URI | undefined {
-		if (!this.productService.extensionsGallery || !this.gallery) {
-			return undefined;
-		}
-
-		return resources.joinPath(URI.parse(this.productService.extensionsGallery.publisherUrl), this.publisher);
-	}
-
-	get publisherDomain(): { link: string; verified: boolean } | undefined {
-		return this.gallery?.publisherDomain;
-	}
-
-	get publisherSponsorLink(): URI | undefined {
-		return this.gallery?.publisherSponsorLink ? URI.parse(this.gallery.publisherSponsorLink) : undefined;
-	}
-
-	get version(): string {
-		return this.local ? this.local.manifest.version : this.latestVersion;
-	}
-
-	get pinned(): boolean {
-		return !!this.local?.pinned;
-	}
-
-	get latestVersion(): string {
-		return this.gallery ? this.gallery.version : this.getManifestFromLocalOrResource()?.version ?? '';
-	}
-
-	get description(): string {
-		return this.gallery ? this.gallery.description : this.getManifestFromLocalOrResource()?.description ?? '';
-	}
-
-	get url(): string | undefined {
-		if (!this.productService.extensionsGallery || !this.gallery) {
-			return undefined;
-		}
-
-		return `${this.productService.extensionsGallery.itemUrl}?itemName=${this.publisher}.${this.name}`;
-	}
-
-	get iconUrl(): string {
-		return this.galleryIconUrl || this.resourceExtensionIconUrl || this.localIconUrl || this.defaultIconUrl;
-	}
-
-	get iconUrlFallback(): string {
-		return this.galleryIconUrlFallback || this.resourceExtensionIconUrl || this.localIconUrl || this.defaultIconUrl;
-	}
-
-	private get localIconUrl(): string | null {
-		if (this.local && this.local.manifest.icon) {
-			return FileAccess.uriToBrowserUri(resources.joinPath(this.local.location, this.local.manifest.icon)).toString(true);
-		}
-		return null;
-	}
-
-	private get resourceExtensionIconUrl(): string | null {
-		if (this.resourceExtension?.manifest.icon) {
-			return FileAccess.uriToBrowserUri(resources.joinPath(this.resourceExtension.location, this.resourceExtension.manifest.icon)).toString(true);
-		}
-		return null;
-	}
-
-	private get galleryIconUrl(): string | null {
-		return this.gallery?.assets.icon ? this.gallery.assets.icon.uri : null;
-	}
-
-	private get galleryIconUrlFallback(): string | null {
-		return this.gallery?.assets.icon ? this.gallery.assets.icon.fallbackUri : null;
-	}
-
-	private get defaultIconUrl(): string {
-		if (this.type === ExtensionType.System && this.local) {
-			if (this.local.manifest && this.local.manifest.contributes) {
-				if (Array.isArray(this.local.manifest.contributes.themes) && this.local.manifest.contributes.themes.length) {
-					return FileAccess.asBrowserUri('vs/workbench/contrib/extensions/browser/media/theme-icon.png').toString(true);
-				}
-				if (Array.isArray(this.local.manifest.contributes.grammars) && this.local.manifest.contributes.grammars.length) {
-					return FileAccess.asBrowserUri('vs/workbench/contrib/extensions/browser/media/language-icon.svg').toString(true);
-				}
-			}
-		}
-		return DefaultIconPath;
-	}
-
-	get repository(): string | undefined {
-		return this.gallery && this.gallery.assets.repository ? this.gallery.assets.repository.uri : undefined;
-	}
-
-	get licenseUrl(): string | undefined {
-		return this.gallery && this.gallery.assets.license ? this.gallery.assets.license.uri : undefined;
-	}
-
-	get supportUrl(): string | undefined {
-		return this.gallery && this.gallery.supportLink ? this.gallery.supportLink : undefined;
-	}
-
-	get state(): ExtensionState {
-		return this.stateProvider(this);
-	}
-
-	public isMalicious: boolean = false;
-	public deprecationInfo: IDeprecationInfo | undefined;
-
-	get installCount(): number | undefined {
-		return this.gallery ? this.gallery.installCount : undefined;
-	}
-
-	get rating(): number | undefined {
-		return this.gallery ? this.gallery.rating : undefined;
-	}
-
-	get ratingCount(): number | undefined {
-		return this.gallery ? this.gallery.ratingCount : undefined;
-	}
-
-	get outdated(): boolean {
-		try {
-			if (!this.gallery || !this.local) {
-				return false;
-			}
-			// Do not allow updating system extensions in stable
-			if (this.type === ExtensionType.System && this.productService.quality === 'stable') {
-				return false;
-			}
-			if (!this.local.preRelease && this.gallery.properties.isPreReleaseVersion) {
-				return false;
-			}
-			if (semver.gt(this.latestVersion, this.version)) {
-				return true;
-			}
-			if (this.outdatedTargetPlatform) {
-				return true;
-			}
-		} catch (error) {
-			/* Ignore */
-		}
-		return false;
-	}
-
-	get outdatedTargetPlatform(): boolean {
-		return !!this.local && !!this.gallery
-			&& ![TargetPlatform.UNDEFINED, TargetPlatform.WEB].includes(this.local.targetPlatform)
-			&& this.gallery.properties.targetPlatform !== TargetPlatform.WEB
-			&& this.local.targetPlatform !== this.gallery.properties.targetPlatform
-			&& semver.eq(this.latestVersion, this.version);
-	}
-
-	get runtimeState(): ExtensionRuntimeState | undefined {
-		return this.runtimeStateProvider(this);
-	}
-
-	get telemetryData(): any {
-		const { local, gallery } = this;
-
-		if (gallery) {
-			return getGalleryExtensionTelemetryData(gallery);
-		} else if (local) {
-			return getLocalExtensionTelemetryData(local);
-		} else {
-			return {};
-		}
-	}
-
-	get preview(): boolean {
-		return this.local?.manifest.preview ?? this.gallery?.preview ?? false;
-	}
-
-	get preRelease(): boolean {
-		return !!this.local?.preRelease;
-	}
-
-	get isPreReleaseVersion(): boolean {
-		if (this.local) {
-			return this.local.isPreReleaseVersion;
-		}
-		return !!this.gallery?.properties.isPreReleaseVersion;
-	}
-
-	get hasPreReleaseVersion(): boolean {
-		return !!this.gallery?.hasPreReleaseVersion || !!this.local?.hasPreReleaseVersion;
-	}
-
-	get hasReleaseVersion(): boolean {
-		return !!this.resourceExtension || !!this.gallery?.hasReleaseVersion;
-	}
-
-	private getLocal(): ILocalExtension | undefined {
-		return this.local && !this.outdated ? this.local : undefined;
-	}
-
-	async getManifest(token: CancellationToken): Promise<IExtensionManifest | null> {
-		const local = this.getLocal();
-		if (local) {
-			return local.manifest;
-		}
-
-		if (this.gallery) {
-			if (this.gallery.assets.manifest) {
-				return this.galleryService.getManifest(this.gallery, token);
-			}
-			this.logService.error(nls.localize('Manifest is not found', "Manifest is not found"), this.identifier.id);
-			return null;
-		}
-
-		if (this.resourceExtension) {
-			return this.resourceExtension.manifest;
-		}
-
-		return null;
-	}
-
-	hasReadme(): boolean {
-		if (this.local && this.local.readmeUrl) {
-			return true;
-		}
-
-		if (this.gallery && this.gallery.assets.readme) {
-			return true;
-		}
-
-		if (this.resourceExtension?.readmeUri) {
-			return true;
-		}
-
-		return this.type === ExtensionType.System;
-	}
-
-	async getReadme(token: CancellationToken): Promise<string> {
-		const local = this.getLocal();
-		if (local?.readmeUrl) {
-			const content = await this.fileService.readFile(local.readmeUrl);
-			return content.value.toString();
-		}
-
-		if (this.gallery) {
-			if (this.gallery.assets.readme) {
-				return this.galleryService.getReadme(this.gallery, token);
-			}
-			this.telemetryService.publicLog('extensions:NotFoundReadMe', this.telemetryData);
-		}
-
-		if (this.type === ExtensionType.System) {
-			return Promise.resolve(`# ${this.displayName || this.name}
-**Notice:** This extension is bundled with Visual Studio Code. It can be disabled but not uninstalled.
-## Features
-${this.description}
-`);
-		}
-
-		if (this.resourceExtension?.readmeUri) {
-			const content = await this.fileService.readFile(this.resourceExtension?.readmeUri);
-			return content.value.toString();
-		}
-
-		return Promise.reject(new Error('not available'));
-	}
-
-	hasChangelog(): boolean {
-		if (this.local && this.local.changelogUrl) {
-			return true;
-		}
-
-		if (this.gallery && this.gallery.assets.changelog) {
-			return true;
-		}
-
-		return this.type === ExtensionType.System;
-	}
-
-	async getChangelog(token: CancellationToken): Promise<string> {
-		const local = this.getLocal();
-		if (local?.changelogUrl) {
-			const content = await this.fileService.readFile(local.changelogUrl);
-			return content.value.toString();
-		}
-
-		if (this.gallery?.assets.changelog) {
-			return this.galleryService.getChangelog(this.gallery, token);
-		}
-
-		if (this.type === ExtensionType.System) {
-			return Promise.resolve(`Please check the [VS Code Release Notes](command:${ShowCurrentReleaseNotesActionId}) for changes to the built-in extensions.`);
-		}
-
-		return Promise.reject(new Error('not available'));
-	}
-
-	get categories(): readonly string[] {
-		const { local, gallery, resourceExtension } = this;
-		if (local && local.manifest.categories && !this.outdated) {
-			return local.manifest.categories;
-		}
-		if (gallery) {
-			return gallery.categories;
-		}
-		if (resourceExtension) {
-			return resourceExtension.manifest.categories ?? [];
-		}
-		return [];
-	}
-
-	get tags(): readonly string[] {
-		const { gallery } = this;
-		if (gallery) {
-			return gallery.tags.filter(tag => !tag.startsWith('_'));
-		}
-		return [];
-	}
-
-	get dependencies(): string[] {
-		const { local, gallery, resourceExtension } = this;
-		if (local && local.manifest.extensionDependencies && !this.outdated) {
-			return local.manifest.extensionDependencies;
-		}
-		if (gallery) {
-			return gallery.properties.dependencies || [];
-		}
-		if (resourceExtension) {
-			return resourceExtension.manifest.extensionDependencies || [];
-		}
-		return [];
-	}
-
-	get extensionPack(): string[] {
-		const { local, gallery, resourceExtension } = this;
-		if (local && local.manifest.extensionPack && !this.outdated) {
-			return local.manifest.extensionPack;
-		}
-		if (gallery) {
-			return gallery.properties.extensionPack || [];
-		}
-		if (resourceExtension) {
-			return resourceExtension.manifest.extensionPack || [];
-		}
-		return [];
-	}
-
-	private getManifestFromLocalOrResource(): IExtensionManifest | null {
-		if (this.local) {
-			return this.local.manifest;
-		}
-		if (this.resourceExtension) {
-			return this.resourceExtension.manifest;
-		}
-		return null;
-	}
-}
-
 const EXTENSIONS_AUTO_UPDATE_KEY = 'extensions.autoUpdate';
-
-class Extensions extends Disposable {
-
-	static updateExtensionFromControlManifest(extension: Extension, extensionsControlManifest: IExtensionsControlManifest): void {
-		extension.isMalicious = extensionsControlManifest.malicious.some(identifier => areSameExtensions(extension.identifier, identifier));
-		extension.deprecationInfo = extensionsControlManifest.deprecated ? extensionsControlManifest.deprecated[extension.identifier.id.toLowerCase()] : undefined;
-	}
-
-	private readonly _onChange = this._register(new Emitter<{ extension: Extension; operation?: InstallOperation } | undefined>());
-	get onChange() { return this._onChange.event; }
-
-	private readonly _onReset = this._register(new Emitter<void>());
-	get onReset() { return this._onReset.event; }
-
-	private installing: Extension[] = [];
-	private uninstalling: Extension[] = [];
-	private installed: Extension[] = [];
-
-	constructor(
-		readonly server: IExtensionManagementServer,
-		private readonly stateProvider: IExtensionStateProvider<ExtensionState>,
-		private readonly runtimeStateProvider: IExtensionStateProvider<ExtensionRuntimeState | undefined>,
-		private readonly isWorkspaceServer: boolean,
-		@IExtensionGalleryService private readonly galleryService: IExtensionGalleryService,
-		@IWorkbenchExtensionEnablementService private readonly extensionEnablementService: IWorkbenchExtensionEnablementService,
-		@IWorkbenchExtensionManagementService private readonly workbenchExtensionManagementService: IWorkbenchExtensionManagementService,
-		@IUserDataProfileService private readonly userDataProfileService: IUserDataProfileService,
-		@ITelemetryService private readonly telemetryService: ITelemetryService,
-		@IInstantiationService private readonly instantiationService: IInstantiationService
-	) {
-		super();
-		this._register(server.extensionManagementService.onInstallExtension(e => this.onInstallExtension(e)));
-		this._register(server.extensionManagementService.onDidInstallExtensions(e => this.onDidInstallExtensions(e)));
-		this._register(server.extensionManagementService.onUninstallExtension(e => this.onUninstallExtension(e.identifier)));
-		this._register(server.extensionManagementService.onDidUninstallExtension(e => this.onDidUninstallExtension(e)));
-		this._register(server.extensionManagementService.onDidUpdateExtensionMetadata(e => this.onDidUpdateExtensionMetadata(e.local)));
-		this._register(server.extensionManagementService.onDidChangeProfile(() => this.reset()));
-		this._register(extensionEnablementService.onEnablementChanged(e => this.onEnablementChanged(e)));
-		this._register(Event.any(this.onChange, this.onReset)(() => this._local = undefined));
-		if (this.isWorkspaceServer) {
-			this._register(this.workbenchExtensionManagementService.onInstallExtension(e => {
-				if (e.workspaceScoped) {
-					this.onInstallExtension(e);
-				}
-			}));
-			this._register(this.workbenchExtensionManagementService.onDidInstallExtensions(e => {
-				const result = e.filter(e => e.workspaceScoped);
-				if (result.length) {
-					this.onDidInstallExtensions(result);
-				}
-			}));
-			this._register(this.workbenchExtensionManagementService.onUninstallExtension(e => {
-				if (e.workspaceScoped) {
-					this.onUninstallExtension(e.identifier);
-				}
-			}));
-			this._register(this.workbenchExtensionManagementService.onDidUninstallExtension(e => {
-				if (e.workspaceScoped) {
-					this.onDidUninstallExtension(e);
-				}
-			}));
-		}
-	}
-
-	private _local: IExtension[] | undefined;
-	get local(): IExtension[] {
-		if (!this._local) {
-			this._local = [];
-			for (const extension of this.installed) {
-				this._local.push(extension);
-			}
-			for (const extension of this.installing) {
-				if (!this.installed.some(installed => areSameExtensions(installed.identifier, extension.identifier))) {
-					this._local.push(extension);
-				}
-			}
-		}
-		return this._local;
-	}
-
-	async queryInstalled(productVersion: IProductVersion): Promise<IExtension[]> {
-		await this.fetchInstalledExtensions(productVersion);
-		this._onChange.fire(undefined);
-		return this.local;
-	}
-
-	async syncInstalledExtensionsWithGallery(galleryExtensions: IGalleryExtension[], productVersion: IProductVersion): Promise<void> {
-		const extensions = await this.mapInstalledExtensionWithCompatibleGalleryExtension(galleryExtensions, productVersion);
-		for (const [extension, gallery] of extensions) {
-			// update metadata of the extension if it does not exist
-			if (extension.local && !extension.local.identifier.uuid) {
-				extension.local = await this.updateMetadata(extension.local, gallery);
-			}
-			if (!extension.gallery || extension.gallery.version !== gallery.version || extension.gallery.properties.targetPlatform !== gallery.properties.targetPlatform) {
-				extension.gallery = gallery;
-				this._onChange.fire({ extension });
-			}
-		}
-	}
-
-	private async mapInstalledExtensionWithCompatibleGalleryExtension(galleryExtensions: IGalleryExtension[], productVersion: IProductVersion): Promise<[Extension, IGalleryExtension][]> {
-		const mappedExtensions = this.mapInstalledExtensionWithGalleryExtension(galleryExtensions);
-		const targetPlatform = await this.server.extensionManagementService.getTargetPlatform();
-		const compatibleGalleryExtensions: IGalleryExtension[] = [];
-		const compatibleGalleryExtensionsToFetch: IExtensionInfo[] = [];
-		await Promise.allSettled(mappedExtensions.map(async ([extension, gallery]) => {
-			if (extension.local) {
-				if (await this.galleryService.isExtensionCompatible(gallery, extension.local.preRelease, targetPlatform, productVersion)) {
-					compatibleGalleryExtensions.push(gallery);
-				} else {
-					compatibleGalleryExtensionsToFetch.push({ ...extension.local.identifier, preRelease: extension.local.preRelease });
-				}
-			}
-		}));
-		if (compatibleGalleryExtensionsToFetch.length) {
-			const result = await this.galleryService.getExtensions(compatibleGalleryExtensionsToFetch, { targetPlatform, compatible: true, queryAllVersions: true, productVersion }, CancellationToken.None);
-			compatibleGalleryExtensions.push(...result);
-		}
-		return this.mapInstalledExtensionWithGalleryExtension(compatibleGalleryExtensions);
-	}
-
-	private mapInstalledExtensionWithGalleryExtension(galleryExtensions: IGalleryExtension[]): [Extension, IGalleryExtension][] {
-		const mappedExtensions: [Extension, IGalleryExtension][] = [];
-		const byUUID = new Map<string, IGalleryExtension>(), byID = new Map<string, IGalleryExtension>();
-		for (const gallery of galleryExtensions) {
-			byUUID.set(gallery.identifier.uuid, gallery);
-			byID.set(gallery.identifier.id.toLowerCase(), gallery);
-		}
-		for (const installed of this.installed) {
-			if (installed.uuid) {
-				const gallery = byUUID.get(installed.uuid);
-				if (gallery) {
-					mappedExtensions.push([installed, gallery]);
-					continue;
-				}
-			}
-			if (installed.local?.source !== 'resource') {
-				const gallery = byID.get(installed.identifier.id.toLowerCase());
-				if (gallery) {
-					mappedExtensions.push([installed, gallery]);
-				}
-			}
-		}
-		return mappedExtensions;
-	}
-
-	private async updateMetadata(localExtension: ILocalExtension, gallery: IGalleryExtension): Promise<ILocalExtension> {
-		let isPreReleaseVersion = false;
-		if (localExtension.manifest.version !== gallery.version) {
-			type GalleryServiceMatchInstalledExtensionClassification = {
-				owner: 'sandy081';
-				comment: 'Report when a request is made to update metadata of an installed extension';
-			};
-			this.telemetryService.publicLog2<{}, GalleryServiceMatchInstalledExtensionClassification>('galleryService:updateMetadata');
-			const galleryWithLocalVersion: IGalleryExtension | undefined = (await this.galleryService.getExtensions([{ ...localExtension.identifier, version: localExtension.manifest.version }], CancellationToken.None))[0];
-			isPreReleaseVersion = !!galleryWithLocalVersion?.properties?.isPreReleaseVersion;
-		}
-		return this.server.extensionManagementService.updateMetadata(localExtension, { id: gallery.identifier.uuid, publisherDisplayName: gallery.publisherDisplayName, publisherId: gallery.publisherId, isPreReleaseVersion }, this.userDataProfileService.currentProfile.extensionsResource);
-	}
-
-	canInstall(galleryExtension: IGalleryExtension): Promise<boolean> {
-		return this.server.extensionManagementService.canInstall(galleryExtension);
-	}
-
-	private onInstallExtension(event: InstallExtensionEvent): void {
-		const { source } = event;
-		if (source && !URI.isUri(source)) {
-			const extension = this.installed.find(e => areSameExtensions(e.identifier, source.identifier))
-				?? this.instantiationService.createInstance(Extension, this.stateProvider, this.runtimeStateProvider, this.server, undefined, source, undefined);
-			this.installing.push(extension);
-			this._onChange.fire({ extension });
-		}
-	}
-
-	private async fetchInstalledExtensions(productVersion?: IProductVersion): Promise<void> {
-		const extensionsControlManifest = await this.server.extensionManagementService.getExtensionsControlManifest();
-		const all = await this.server.extensionManagementService.getInstalled(undefined, undefined, productVersion);
-		if (this.isWorkspaceServer) {
-			all.push(...await this.workbenchExtensionManagementService.getInstalledWorkspaceExtensions(true));
-		}
-
-		// dedup workspace, user and system extensions by giving priority to workspace first and then to user extension.
-		const installed = groupByExtension(all, r => r.identifier).reduce((result, extensions) => {
-			if (extensions.length === 1) {
-				result.push(extensions[0]);
-			} else {
-				let workspaceExtension: ILocalExtension | undefined,
-					userExtension: ILocalExtension | undefined,
-					systemExtension: ILocalExtension | undefined;
-				for (const extension of extensions) {
-					if (extension.isWorkspaceScoped) {
-						workspaceExtension = extension;
-					} else if (extension.type === ExtensionType.User) {
-						userExtension = extension;
-					} else {
-						systemExtension = extension;
-					}
-				}
-				const extension = workspaceExtension ?? userExtension ?? systemExtension;
-				if (extension) {
-					result.push(extension);
-				}
-			}
-			return result;
-		}, []);
-
-		const byId = index(this.installed, e => e.local ? e.local.identifier.id : e.identifier.id);
-		this.installed = installed.map(local => {
-			const extension = byId[local.identifier.id] || this.instantiationService.createInstance(Extension, this.stateProvider, this.runtimeStateProvider, this.server, local, undefined, undefined);
-			extension.local = local;
-			extension.enablementState = this.extensionEnablementService.getEnablementState(local);
-			Extensions.updateExtensionFromControlManifest(extension, extensionsControlManifest);
-			return extension;
-		});
-	}
-
-	private async reset(): Promise<void> {
-		this.installed = [];
-		this.installing = [];
-		this.uninstalling = [];
-		await this.fetchInstalledExtensions();
-		this._onReset.fire();
-	}
-
-	private async onDidInstallExtensions(results: readonly InstallExtensionResult[]): Promise<void> {
-		for (const event of results) {
-			const { local, source } = event;
-			const gallery = source && !URI.isUri(source) ? source : undefined;
-			const location = source && URI.isUri(source) ? source : undefined;
-			const installingExtension = gallery ? this.installing.filter(e => areSameExtensions(e.identifier, gallery.identifier))[0] : null;
-			this.installing = installingExtension ? this.installing.filter(e => e !== installingExtension) : this.installing;
-
-			let extension: Extension | undefined = installingExtension ? installingExtension
-				: (location || local) ? this.instantiationService.createInstance(Extension, this.stateProvider, this.runtimeStateProvider, this.server, local, undefined, undefined)
-					: undefined;
-			if (extension) {
-				if (local) {
-					const installed = this.installed.filter(e => areSameExtensions(e.identifier, extension!.identifier))[0];
-					if (installed) {
-						extension = installed;
-					} else {
-						this.installed.push(extension);
-					}
-					extension.local = local;
-					if (!extension.gallery) {
-						extension.gallery = gallery;
-					}
-					Extensions.updateExtensionFromControlManifest(extension, await this.server.extensionManagementService.getExtensionsControlManifest());
-					extension.enablementState = this.extensionEnablementService.getEnablementState(local);
-				}
-			}
-			this._onChange.fire(!local || !extension ? undefined : { extension, operation: event.operation });
-			if (extension && extension.local && !extension.gallery && extension.local.source !== 'resource') {
-				await this.syncInstalledExtensionWithGallery(extension);
-			}
-		}
-	}
-
-	private async onDidUpdateExtensionMetadata(local: ILocalExtension): Promise<void> {
-		const extension = this.installed.find(e => areSameExtensions(e.identifier, local.identifier));
-		if (extension?.local) {
-			const hasChanged = extension.local.pinned !== local.pinned
-				|| extension.local.preRelease !== local.preRelease;
-			extension.local = local;
-			if (hasChanged) {
-				this._onChange.fire({ extension });
-			}
-		}
-	}
-
-	private async syncInstalledExtensionWithGallery(extension: Extension): Promise<void> {
-		if (!this.galleryService.isEnabled()) {
-			return;
-		}
-		type GalleryServiceMatchInstalledExtensionClassification = {
-			owner: 'sandy081';
-			comment: 'Report when a request is made to match installed extension with gallery';
-		};
-		this.telemetryService.publicLog2<{}, GalleryServiceMatchInstalledExtensionClassification>('galleryService:matchInstalledExtension');
-		const [compatible] = await this.galleryService.getExtensions([{ ...extension.identifier, preRelease: extension.local?.preRelease }], { compatible: true, targetPlatform: await this.server.extensionManagementService.getTargetPlatform() }, CancellationToken.None);
-		if (compatible) {
-			extension.gallery = compatible;
-			this._onChange.fire({ extension });
-		}
-	}
-
-	private onUninstallExtension(identifier: IExtensionIdentifier): void {
-		const extension = this.installed.filter(e => areSameExtensions(e.identifier, identifier))[0];
-		if (extension) {
-			const uninstalling = this.uninstalling.filter(e => areSameExtensions(e.identifier, identifier))[0] || extension;
-			this.uninstalling = [uninstalling, ...this.uninstalling.filter(e => !areSameExtensions(e.identifier, identifier))];
-			this._onChange.fire(uninstalling ? { extension: uninstalling } : undefined);
-		}
-	}
-
-	private onDidUninstallExtension({ identifier, error }: DidUninstallExtensionEvent): void {
-		const uninstalled = this.uninstalling.find(e => areSameExtensions(e.identifier, identifier)) || this.installed.find(e => areSameExtensions(e.identifier, identifier));
-		this.uninstalling = this.uninstalling.filter(e => !areSameExtensions(e.identifier, identifier));
-		if (!error) {
-			this.installed = this.installed.filter(e => !areSameExtensions(e.identifier, identifier));
-		}
-		if (uninstalled) {
-			this._onChange.fire({ extension: uninstalled });
-		}
-	}
-
-	private onEnablementChanged(platformExtensions: readonly IPlatformExtension[]) {
-		const extensions = this.local.filter(e => platformExtensions.some(p => areSameExtensions(e.identifier, p.identifier)));
-		for (const extension of extensions) {
-			if (extension.local) {
-				const enablementState = this.extensionEnablementService.getEnablementState(extension.local);
-				if (enablementState !== extension.enablementState) {
-					(extension as Extension).enablementState = enablementState;
-					this._onChange.fire({ extension: extension as Extension });
-				}
-			}
-		}
-	}
-
-	getExtensionState(extension: Extension): ExtensionState {
-		if (extension.gallery && this.installing.some(e => !!e.gallery && areSameExtensions(e.gallery.identifier, extension.gallery!.identifier))) {
-			return ExtensionState.Installing;
-		}
-		if (this.uninstalling.some(e => areSameExtensions(e.identifier, extension.identifier))) {
-			return ExtensionState.Uninstalling;
-		}
-		const local = this.installed.filter(e => e === extension || (e.gallery && extension.gallery && areSameExtensions(e.gallery.identifier, extension.gallery.identifier)))[0];
-		return local ? ExtensionState.Installed : ExtensionState.Uninstalled;
-	}
-}
 
 export class ExtensionsWorkbenchService extends Disposable implements IExtensionsWorkbenchService, IURLHandler {
