+++ src/vs/workbench/contrib/extensions/browser/newFile.ts
@@ -1,0 +1,1 @@
+
